refactor: adjust to AC SfDropdown component
