--- packages/shared/styles/components/organisms/SfDropdown.scss
@@ -1,57 +1,0 @@
-@import "../../helpers";
-
-.sf-dropdown {
-  &__list {
-    margin: var(--dropdown-list-margin, 0);
-    box-shadow: var(--dropdown-list-box-shadow, none);
-    z-index: var(--dropdown-z-index, 1);
-    box-sizing: border-box;
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    width: 100%;
-    &::before {
-      content: "";
-      display: var(--dropdown-before-display, block);
-      background-color: var(--dropdown-before-background-color, var(--c-body));
-      width: 3.75rem;
-      height: 3px;
-      margin: 0 auto var(--spacer-2xs);
-      border-radius: 4px;
-    }
-  }
-  &__button {
-    padding: var(--spacer-base) 0;
-    width: 100%;
-    display: block;
-    &--title {
-      --button-font-weight: var(--font-normal);
-      --button-font-size: var(--font-sm);
-      --button-color: var(--c-primary);
-      --button-background: var(--c-white);
-      --button-text-transform: capitalize;
-    }
-    &--cancell {
-      --button-color: var(--c-text-muted);
-      --button-background: var(--c-light);
-    }
-  }
-  @include for-desktop {
-    position: relative;
-    &__list {
-      --dropdown-list-box-shadow: 0px 4px 11px rgba(29, 31, 34, 0.1);
-      position: relative;
-      width: var(--dropdown-list-width, 300px);
-      --dropdown-before-display: none;
-      --dropdown-list-margin: var(--spacer-sm) 0 0 0;
-    }
-    &--up {
-      .sf-dropdown__list {
-        position: absolute;
-        bottom: 100%;
-        left: 0;
-        margin: 0 0 var(--spacer-xs) 0;
-      }
-    }
-  }
-}
