--- packages/vue/src/components/organisms/SfDropdown/_internal/SfDropdownItem.vue
@@ -1,10 +1,0 @@
-<template>
-  <div class="sf-dropdown-item">
-    <slot />
-  </div>
-</template>
-<script>
-export default {
-  name: "SfDropdownItem"
-};
-</script>
